move overlay-checks to an overlayutils package
This allows using the utilities without importing the whole
snapshotter.
